Editor loads with new editor interface
