+++ src/editor/commands/SetPositionCommand.js
@@ -1,5 +1,5 @@
 import Command from "./Command";
-import { TransformSpace } from "../Editor2";
+import { TransformSpace } from "../Editor";
 
 export default class SetPositionCommand extends Command {
   constructor(editor, object, position, space) {
