+++ src/editor/commands/TranslateCommand.js
@@ -1,5 +1,5 @@
 import Command from "./Command";
-import { TransformSpace } from "../Editor2";
+import { TransformSpace } from "../Editor";
 
 export default class TranslateCommand extends Command {
   constructor(editor, object, translation, space) {
