+++ src/editor/nodes/ModelNode.js
@@ -149,7 +149,7 @@ export default class ModelNode extends EditorNodeMixin(Model) {
         this.scaleToFit = false;
       }
 
-      this.editor.signals.objectChanged.dispatch(this);
+      this.editor.emit("objectsChanged", [this]);
 
       if (files) {
         // Revoke any object urls from the SketchfabZipLoader.
