+++ src/ui/properties/VideoNodeEditor.js
@@ -27,7 +27,7 @@ export default class VideoNodeEditor extends Component {
 
   constructor(props) {
     super(props);
-    const createPropSetter = propName => value => this.props.editor.setNodeProperty(this.props.node, propName, value);
+    const createPropSetter = propName => value => this.props.editor.setProperty(this.props.node, propName, value);
     this.onChangeSrc = createPropSetter("src");
     this.onChangeProjection = createPropSetter("projection");
     this.onChangeControls = createPropSetter("controls");
