--- src/ui/viewport/RendererStats.scss
@@ -1,31 +1,0 @@
-@import "../styles/theme";
-
-:local(.rendererStats) {
-  position: absolute;
-  left: 0px;
-  bottom: 0px;
-  background-color: change_color($panel, $alpha: 0.4);
-  width: 100%;
-  height: 1.5em;
-  line-height: 1.5em;
-  margin: 0;
-  padding: 0 0.5em 0 0.5em;
-
-  > dt, dd {
-    display: inline;
-  }
-
-  > dt {
-    color: darken($text, 10%);
-    margin: 0 0.3em 0 0;
-  }
-
-  > dd {
-    color: $text;
-    margin: 0 0.75em 0 0;
-  }
-
-  > dt::after {
-    content: ":";
-  }
-}
